requirements.txt
streamlit>=1.30.0
openai>=1.10.0
requests>=2.31.0
yt-dlp>=2024.1.0
imageio-ffmpeg>=0.4.9
python-dotenv>=1.0.0
anthropic>=0.40.0
apify-client>=1.7.0
gspread>=5.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
